Use correct return type for contentLength

+++ CocoaHTTPServer/Core/Responses/HTTPAsyncFileResponse.m
@@ -274,7 +274,7 @@ static const int httpLogLevel = HTTP_LOG_LEVEL_WARN; // | HTTP_LOG_FLAG_TRACE;
 	return [self openFileAndSetupReadSource];
 }	
 
-- (UInt64)contentLength
+- (NSUInteger)contentLength
 {
 	HTTPLogTrace2(@"%@[%p]: contentLength - %llu", THIS_FILE, self, fileLength);
